Remove wildcard condition from parameters
AFAICT wildcard parameters are specified by named parameters, eg
"theta".

Address PR comments. Squash.

Move "concretization" into the INSTANTIATE-BINDING function.

INSTANTIATE-BINDING is meant to take a GATE-BINDING object which
specifies for example the operator, parameters, and arguments that a
compiler binding may act upon. It is not unthinkable that we might
want to instantiate a binding whose arguments are wildcards -- i.e. a
compiler who matches against any qubit. Previously, function would
raise an error if that were the case; now, instead, it fills-in
wildcard arguments with unique qubit indices.